1、）修改分页获取消息界面点赞和申请加微信消息列表接口 修改返回点赞消息数据： 字段，content1：最新一条点赞消息 新增返回评论消息数据： 字段，comments：最新评论消息 字段，content2：最新一条评论消息内容 字段，count2：未读评论消息数量 详情见文档 2、）新增分页获取评论消息列表接口 字段见文档 3、）新增删除评论消息记录接口 字段见文档

+++ find-common/src/main/java/com/potato369/find/common/vo/LikesMessageVO.java
@@ -22,5 +22,5 @@ public class LikesMessageVO {
     //未读点赞消息总条数
     @JSONField(name = "count1")
     @JsonProperty(value = "count1")
-    private Long count;
+    private Integer count;
 }
